+++ app/api/webhook/route.ts
@@ -1,8 +1,9 @@
+/* eslint-disable camelcase */
 import { Webhook } from "svix";
 import { headers } from "next/headers";
-import { createUser, deleteUser, updateUser } from "@/lib/actions/user.action";
 import { UserJSON, WebhookEvent, clerkClient } from "@clerk/nextjs/server";
 import { NextResponse } from "next/server";
+import { createUser, deleteUser, updateUser } from "@/lib/actions/userActions";
 
 export async function POST(req: Request) {
   const webhookSecret = process.env.CLERK_WEBHOOK_SECRET;
